requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Bcrypt
Flask-Mail
marshmallow
python-dotenv
PyMySQL
cryptography
Pillow
cloudinary
requests
python-dateutil
validators
pytest
pytest-flask
flask-bcrypt
paypalrestsdk
Flask-Session
langchain
langchain-openai
langchain-core
langchain-community